+++ project/step3.cpp
@@ -8,6 +8,8 @@
 #include <LBFGSB.h>
 #include <sstream>
 
+template class VolSurfBuilder<CubicSmile>;
+
 std::string convertUnixMSToISO8601(uint64_t timestamp) {
     std::time_t time = static_cast<std::time_t>(timestamp / 1000);
     std::tm* dateTime = std::gmtime(&time);
